requirements.txt
pandas
google-cloud-firestore
google-cloud-storage
python-dotenv 
qrcode
Pillow
Flask
gunicorn